deep form signal enhancements
