+++ projects/ngx-form-signal/src/public-api.ts
@@ -2,7 +2,9 @@
  * Public API Surface of ngx-form-signal
  */
 
+export * from './lib/deep-form-signal';
 export * from './lib/form-signal';
+export * from './lib/types/deep-form-signal-type';
 export * from './lib/types/form-signal-options';
 export * from './lib/types/form-signal-type';
 export * from './lib/types/form-type';
