+++ projects/test-app/src/app/deep-form-signal-example/deep-form-signal-example.component.ts
@@ -6,7 +6,10 @@ import {
    ReactiveFormsModule,
    Validators,
 } from '@angular/forms';
-import { deepFormSignal } from '../../../../ngx-form-signal/src/public-api';
+import {
+   deepFormSignal,
+   formSignal,
+} from '../../../../ngx-form-signal/src/public-api';
 
 @Component({
    selector: 'deep-form-signal-example',
@@ -29,40 +32,41 @@ export class DeepFormSignalExampleComponent {
       }),
    });
 
-   readonly formSignal = deepFormSignal(this.form);
+   readonly formSignal = formSignal(this.form);
+   readonly deepFormSignal = deepFormSignal(this.form);
 
    readonly preview = computed(() => {
       return { ...this.formSignal(), subscriptions: null };
    });
 
    readonly addressEffect = effect(() => {
-      console.log('address:', this.formSignal.children.address.value());
+      console.log('address:', this.deepFormSignal.children.address.value());
    });
    readonly addressStateEffect = effect(() => {
       console.log(
          'address state:',
-         this.formSignal.children.address.children.state.value()
+         this.deepFormSignal.children.address.children.state.value()
       );
    });
    readonly addressStreetEffect = effect(() => {
       console.log(
          'address street:',
-         this.formSignal.children.address.children.street.value()
+         this.deepFormSignal.children.address.children.street.value()
       );
    });
    readonly addressZipEffect = effect(() => {
       console.log(
          'address zip:',
-         this.formSignal.children.address.children.zip.value()
+         this.deepFormSignal.children.address.children.zip.value()
       );
    });
    readonly emailEffect = effect(() => {
-      console.log('email:', this.formSignal.children.email.value());
+      console.log('email:', this.deepFormSignal.children.email.value());
    });
    readonly messageEffect = effect(() => {
-      console.log('message:', this.formSignal.children.message.value());
+      console.log('message:', this.deepFormSignal.children.message.value());
    });
    readonly nameEffect = effect(() => {
-      console.log('name:', this.formSignal.children.name.value());
+      console.log('name:', this.deepFormSignal.children.name.value());
    });
 }
